% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histof.R
\name{abo}
\alias{abo}
\title{test for ABO compatibility}
\usage{
abo(cABO = "A", dABO = "A", iso = TRUE)
}
\arguments{
\item{cABO}{A character from (\code{env$valid.blood.groups})}

\item{dABO}{A character from (\code{env$valid.blood.groups})}

\item{iso}{A logical value T/F}
}
\value{
A logical value T/F
}
\description{
ABO compatibility test between donor and candidate
}
\examples{
abo(cABO = 'A', dABO = 'A', iso = TRUE)
}
