% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk.R
\name{abo_uk}
\alias{abo_uk}
\title{ABO compatibility on UK's algorithm}
\usage{
abo_uk(dABO = "A", cABO = "A", tier = "B")
}
\arguments{
\item{dABO}{A character value for donor's ABO blood group (O, A, B, AB).}

\item{cABO}{A character value for candidate's ABO blood group (O, A, B, AB).}

\item{tier}{A character value for UK transplant candidate's TIER classification (A, B)}
}
\value{
A logical value T/F
}
\description{
ABO compatibility test between donor and candidate according to
TIER classification
}
\examples{
abo_uk(dABO = "A", cABO = "A", tier = "B")
}
