% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ABOpt}
\alias{ABOpt}
\title{ABO relative frequencies.}
\format{
A data frame with 4 rows and 2 variables:
\describe{
\item{abo}{ABO froup}
\item{freq}{relative frequencies}
...
}
}
\source{
\url{http://www.ipst.pt/files/IPST/INFORMACAO_DOCUMENTACAO/AB0_29_2007_pag_5a17.pdf}
}
\usage{
ABOpt
}
\description{
A dataset containing ABO relative frequencies from portuguese blood donors used for MMP computation.
}
\keyword{datasets}
