% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk.R
\name{age_diff}
\alias{age_diff}
\title{Donor recipient age difference}
\usage{
age_diff(donor.age = 60, candidate.age = 50)
}
\arguments{
\item{donor.age}{A numeric value with donor's age.}

\item{candidate.age}{A numeric value with candidate's age.}
}
\value{
A numeric value.
}
\description{
computes punctuation according to donor-recipient age difference
}
\examples{
age_diff(donor.age = 60, candidate.age = 50)
}
