% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cabs}
\alias{cabs}
\title{Transplant candidates' HLA antibodies.}
\format{
A data frame with 752 rows and 2 variables:
\describe{
\item{ID}{Candidates' identifier}
\item{abs}{Candidates HLA antibodies for which they are himunezed to}
...
}
}
\source{
\url{https://balima.shinyapps.io/kars/}
}
\usage{
cabs
}
\description{
A dataset containing HLA antibodies for a group of transplant candidates.
}
\keyword{datasets}
