% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{candidates}
\alias{candidates}
\title{Waiting list with transplant candidates' information.}
\format{
A data frame with 500 rows and 12 variables:
\describe{
\item{ID}{Candidates' identifier}
\item{bg}{Candidates' blood group}
\item{A1}{HLA-A allele 1}
\item{A2}{HLA-A allele 2}
\item{B1}{HLA-B allele 1}
\item{B2}{HLA-B allele 2}
\item{DR1}{HLA-DR allele 1}
\item{DR2}{HLA-DR allele 2}
\item{age}{Candidates' age in years}
\item{dialysis}{Candidates' time on dialysis in months}
\item{cPRA}{Candidates' cPRA percentage}
\item{urgent}{clinical urgency 0/1}
...
}
}
\source{
\url{https://balima.shinyapps.io/kars/}
}
\usage{
candidates
}
\description{
A dataset containing demographics and medical information for a group of transplant candidates.
}
\keyword{datasets}
