% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{candidates.uk}
\alias{candidates.uk}
\title{Waiting list with UK transplant candidates' information.}
\format{
A data frame with 500 rows and 15 variables:
\describe{
\item{ID}{Candidates' identifier}
\item{bg}{Candidates' blood group}
\item{A1}{HLA-A allele 1}
\item{A2}{HLA-A allele 2}
\item{B1}{HLA-B allele 1}
\item{B2}{HLA-B allele 2}
\item{DR1}{HLA-DR allele 1}
\item{DR2}{HLA-DR allele 2}
\item{age}{Candidates' age in years}
\item{dialysis}{Candidates' time on dialysis in months}
\item{cPRA}{Candidates' cPRA percentage}
\item{Tier}{Candidates are ranked on two Tiers (A and B)}
\item{MS}{Candidates' MAtch Score defined as HLA transplantability}
\item{RRI}{A recipient risk score (RRI) is calculated, for each eligible patient using 4 risk factors: age, dialysis at registration, time on dialysis, diabetic status}
\item{urgent}{clinical urgency 0/1}
...
}
}
\source{
\url{https://nhsbtdbe.blob.core.windows.net/umbraco-assets-corp/16915/kidney-allocation-policy-pol186.pdf}
}
\usage{
candidates.uk
}
\description{
A dataset containing demographics and medical information (including factors used in UK transplant) for a group of transplant candidates.
}
\keyword{datasets}
