% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{donors}
\alias{donors}
\title{Kidney donors' information.}
\format{
A data frame with 70 rows and 9 variables:
\describe{
\item{ID}{Donors' identifier}
\item{bg}{Donors' blood group}
\item{A1}{HLA-A allele 1}
\item{A2}{HLA-A allele 2}
\item{B1}{HLA-B allele 1}
\item{B2}{HLA-B allele 2}
\item{DR1}{HLA-DR allele 1}
\item{DR2}{HLA-DR allele 2}
\item{age}{Donors' age in years}
...
}
}
\source{
\url{https://balima.shinyapps.io/kars/}
}
\usage{
donors
}
\description{
A dataset containing demographics and medical information for a pool of donors.
}
\keyword{datasets}
