% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{donors.uk}
\alias{donors.uk}
\title{Kidney donors' information from UK transplant.}
\format{
A data frame with 70 rows and 10 variables:
\describe{
\item{ID}{Donors' identifier}
\item{bg}{Donors' blood group}
\item{A1}{HLA-A allele 1}
\item{A2}{HLA-A allele 2}
\item{B1}{HLA-B allele 1}
\item{B2}{HLA-B allele 2}
\item{DR1}{HLA-DR allele 1}
\item{DR2}{HLA-DR allele 2}
\item{age}{Donors' age in years}
\item{DRI}{A donor risk score (DRI) is calculated for each donor on offer using 7 risk factors: age, history of hypertension; gender; CMV+, eGFR, days in hospital}
...
}
}
\source{
\url{https://nhsbtdbe.blob.core.windows.net/umbraco-assets-corp/16915/kidney-allocation-policy-pol186.pdf}
}
\usage{
donors.uk
}
\description{
A dataset containing demographics and medical information (including factors used in UK transplant) for a pool of donors.
}
\keyword{datasets}
