% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dst}
\alias{dst}
\title{Grantham distance metric.}
\format{
a 20x20 matrix
\describe{
\item{rows}{amino-acid}
\item{columns}{amino-acid}
}
}
\source{
\url{https://www.science.org/doi/10.1126/science.185.4154.862}
}
\usage{
dst
}
\description{
The Grantham score attempts to predict the distance between two amino-acids, in an evolutionary sense.
A lower Grantham score reflects less evolutionary distance.
A higher Grantham score reflects a greater evolutionary distance.
}
\keyword{datasets}
