% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et.R
\name{et}
\alias{et}
\title{resume function for ET algorithm punctuation}
\usage{
et(
  iso = TRUE,
  dABO = "A",
  dA = c("1", "2"),
  dB = c("15", "44"),
  dDR = c("1", "4"),
  donor.age = 65,
  data = candidates,
  month = 2.78,
  mm0 = 400,
  mm1 = 333.33,
  mm2 = 266.67,
  mm3 = 200,
  mm4 = 133.33,
  mm5 = 66.67,
  mm6 = 0,
  df.abs = cabs,
  hlaA = hlaApt,
  hlaB = hlaBpt,
  hlaDR = hlaDRpt,
  abo_freq = ABOpt,
  n = 2,
  check.validity = TRUE
)
}
\arguments{
\item{iso}{A logical value for isogroupal compatibility.}

\item{dABO}{A character value with ABO blood group (\code{env$valid.blood.groups}).}

\item{dA}{donor's HLA-A typing.}

\item{dB}{donor's HLA-B typing.}

\item{dDR}{donor's HLA-DR typing.}

\item{donor.age}{A numeric value with donor's age.}

\item{data}{A data frame containing demographics and medical information for
a group of waitlisted transplant candidates.}

\item{month}{A numeric value with the punctuation for each month
(between env$month.points.minimum and env$month.points.maximum)}

\item{mm0}{A numeric value with points for 0 HLA mm on ETKAS points table}

\item{mm1}{A numeric value with points for 1 HLA mm on ETKAS points table}

\item{mm2}{A numeric value with points for 2 HLA mm on ETKAS points table}

\item{mm3}{A numeric value with points for 3 HLA mm on ETKAS points table}

\item{mm4}{A numeric value with points for 4 HLA mm on ETKAS points table}

\item{mm5}{A numeric value with points for 5 HLA mm on ETKAS points table}

\item{mm6}{A numeric value with points for 6 HLA mm on ETKAS points table}

\item{df.abs}{A data frame with candidates' antibodies.}

\item{hlaA}{A data frame with HLA-A allele frequencies}

\item{hlaB}{A data frame with HLA-B allele frequencies}

\item{hlaDR}{A data frame with HLA-DR allele frequencies}

\item{abo_freq}{A data frame with ABO blood group frequencies}

\item{n}{A positive integer to slice the first candidates.}

\item{check.validity}{Logical to decide whether to validate input.}
}
\value{
An ordered data frame with columns cPRA, HI, pointsET, SP, AM
and 'mmHLA'.
}
\description{
Ordering of waitlisted candidates for a given donor and
according to ETKAS algorithm.
}
\examples{
et(iso = TRUE, dABO = "A",
dA = c("1","2"), dB = c("15","44"), dDR = c("1","4"),
donor.age = 65,
data = candidates, month = 2.78,
mm0 = 400, mm1 = 333.33, mm2 = 266.67, mm3 = 200,
mm4 = 133.33, mm5 = 66.67, mm6 = 0,
df.abs = cabs,
hlaA = hlaApt, hlaB = hlaBpt, hlaDR = hlaDRpt,
abo_freq = ABOpt,
n = 2,
check.validity = TRUE)
}
