% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et.R
\name{et_mmHLA}
\alias{et_mmHLA}
\title{ET points for mmHLA}
\usage{
et_mmHLA(
  dA = c("01", "02"),
  dB = c("03", "05"),
  dDR = c("04", "06"),
  cA = c("01", "02"),
  cB = c("03", "05"),
  cDR = c("04", "06"),
  mm0 = 400,
  mm1 = 333.33,
  mm2 = 266.67,
  mm3 = 200,
  mm4 = 133.33,
  mm5 = 66.67,
  mm6 = 0
)
}
\arguments{
\item{dA}{A two elements character vector donor's HLA-A typing.}

\item{dB}{A two elements character vector donor's HLA-B typing.}

\item{dDR}{A two elements character vector donor's HLA-DR typing.}

\item{cA}{A two elements character vector candidate's HLA-A typing.}

\item{cB}{A two elements character vector candidate's HLA-B typing.}

\item{cDR}{A two elements character vector candidate's HLA-DR typing.}

\item{mm0}{A numeric value with points for 0 HLA mm on ETKAS points table.}

\item{mm1}{A numeric value with points for 1 HLA mm on ETKAS points table.}

\item{mm2}{A numeric value with points for 2 HLA mm on ETKAS points table.}

\item{mm3}{A numeric value with points for 3 HLA mm on ETKAS points table.}

\item{mm4}{A numeric value with points for 4 HLA mm on ETKAS points table.}

\item{mm5}{A numeric value with points for 5 HLA mm on ETKAS points table.}

\item{mm6}{A numeric value with points for 6 HLA mm on ETKAS points table.}
}
\value{
\code{mmA} number of HLA-A mismatches between \code{dA} and \code{cA};
\code{mmB} number of HLA-B mismatches between \code{dB} and \code{cB};
\code{mmDR} number of HLA-DR mismatches between \code{dA}DR and \code{cDR};
\code{mmHLA} as the sum of \code{mmA} + \code{mmB} + \code{mmDR},
and \code{ptsHLA} corresponding to the punctuaction given as argument for number of \code{mmHLA}.
}
\description{
Computes HLA mismatches (mm) and the respective punctuation within ET
Kidney allocation system
}
\examples{
et_mmHLA(dA = c("01","02"), dB = c("03","05"), dDR = c("04","06"),
cA = c("01","02"), cB = c("03","05"), cDR = c("04","06"),
mm0 = 400, mm1 = 333.33, mm2 = 266.67, mm3 = 200,
mm4 = 133.33, mm5 = 66.67, mm6 = 0)
}
