% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hlaBet}
\alias{hlaBet}
\title{HLA-B relative frequencies.}
\format{
A data frame with 57 rows and 3 variables:
\describe{
\item{B}{HLA-B allele}
\item{n}{allele counts}
\item{freq}{allele frequencies}
...
}
}
\source{
\url{https://www.eurotransplant.org/wp-content/uploads/2020/01/H4-Kidney.pdf}
}
\usage{
hlaBet
}
\description{
A dataset containing HLA-B allele relative frequencies from EuroTransplant donors used for MMP computation.
}
\keyword{datasets}
