% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hlaBpt}
\alias{hlaBpt}
\title{HLA-B relative frequencies.}
\format{
A data frame with 34 rows and 3 variables:
\describe{
\item{B}{HLA-B allele}
\item{n}{allele counts}
\item{freq}{allele frequencies}
...
}
}
\source{
\url{https://www.slideshare.net/balima78/lima-2013}
}
\usage{
hlaBpt
}
\description{
A dataset containing HLA-B allele relative frequencies from portuguese donors used for MMP computation.
}
\keyword{datasets}
