% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hlaDRet}
\alias{hlaDRet}
\title{HLA-DR relative frequencies.}
\format{
A data frame with 18 rows and 3 variables:
\describe{
\item{DR}{HLA-DR allele}
\item{n}{allele counts}
\item{freq}{allele frequencies}
...
}
}
\source{
\url{https://www.eurotransplant.org/wp-content/uploads/2020/01/H4-Kidney.pdf}
}
\usage{
hlaDRet
}
\description{
A dataset containing HLA-DR allele relative frequencies from EuroTransplant donors used for MMP computation.
}
\keyword{datasets}
