% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hlaDRpt}
\alias{hlaDRpt}
\title{HLA-DR relative frequencies.}
\format{
A data frame with 13 rows and 3 variables:
\describe{
\item{DR}{HLA-DR allele}
\item{n}{allele counts}
\item{freq}{allele frequencies}
...
}
}
\source{
\url{https://www.slideshare.net/balima78/lima-2013}
}
\usage{
hlaDRpt
}
\description{
A dataset containing HLA-DR allele relative frequencies from portuguese donors used for MMP computation.
}
\keyword{datasets}
