% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lima.R
\name{lima1}
\alias{lima1}
\title{Candidates' ordering according to Lima's algorithm}
\usage{
lima1(
  iso = TRUE,
  dABO = "O",
  dA = c("1", "2"),
  dB = c("15", "44"),
  dDR = c("1", "4"),
  dage = 60,
  df.abs = cabs,
  data = candidates,
  n = 2,
  ...
)
}
\arguments{
\item{iso}{A logical value for isogroupal compatibility.}

\item{dABO}{A character value with ABO blood group.}

\item{dA}{donor's HLA-A typing.}

\item{dB}{donor's HLA-B typing.}

\item{dDR}{donor's HLA-DR typing.}

\item{dage}{A numeric value with donor's age.}

\item{df.abs}{A data frame with candidates' antibodies.}

\item{data}{A data frame containing demographics and medical information for a group of waitlisted transplant candidates.}

\item{n}{A positive integer to slice the first candidates.}
}
\value{
An ordered data frame with a column 'cp' (color priority), 'sp', 'hi' and 'mmHLA'.
}
\description{
Ordering of waitlisted candidates for a given donor and according to to Lima's algorithm.
}
\examples{
lima1(iso = TRUE, dABO = "A", dA = c("1","2"), dB = c("15","44"), dDR = c("1","4"), dage = 65,  data = candidates,  df.abs = abs, n = dim(data)[1])
}
