% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histof.R
\name{mmHLA}
\alias{mmHLA}
\title{number of HLA mismatches}
\usage{
mmHLA(
  dA = c("1", "2"),
  dB = c("5", "7"),
  dDR = c("1", "4"),
  cA = c("1", "2"),
  cB = c("3", "15"),
  cDR = c("4", "7")
)
}
\arguments{
\item{dA}{A two elements character vector with donor's HLA-A typing}

\item{dB}{A two elements character vector with donor's HLA-B typing}

\item{dDR}{A two elements character vector with donor's HLA-DR typing}

\item{cA}{A two elements character vector with candidate's HLA-A typing}

\item{cB}{A two elements character vector with candidate's HLA-B typing}

\item{cDR}{A two elements character vector with candidate's HLA-DR typing}
}
\value{
\code{mmA} number of HLA-A mismatches between \code{dA} and \code{cA};
\code{mmB} number of HLA-B mismatches between \code{dB} and \code{cB};
\code{mmDR} number of HLA-DR mismatches between \code{dA}DR and \code{cDR};
and \code{mmHLA} as the sum of \code{mmA} + \code{mmB} + \code{mmDR}
}
\description{
Computes the number of HLA mismatches between donor and candidate
}
\examples{
mmHLA(dA = c('1','2'), dB = c('5','7'), dDR = c('1','4'),
cA = c('1','2'), cB = c('03','15'), cDR = c('04','07'))
}
