% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.R
\name{pt}
\alias{pt}
\title{Matching punctuation' according to 2007 PT's algorithm}
\usage{
pt(
  iso = TRUE,
  dABO = "O",
  dA = c("1", "2"),
  dB = c("15", "44"),
  dDR = c("1", "4"),
  donor.age = 65,
  df.abs = cabs,
  data = candidates,
  points.80 = 8,
  points.50 = 4,
  points.dialysis = 0.1,
  points.age = 4,
  n = 2,
  check.validity = TRUE
)
}
\arguments{
\item{iso}{A logical value for isogroupal compatibility.}

\item{dABO}{A character value with ABO blood group (\code{env$valid.blood.groups}).}

\item{dA}{donor's HLA-A typing.}

\item{dB}{donor's HLA-B typing.}

\item{dDR}{donor's HLA-DR typing.}

\item{donor.age}{A numeric value with donor's age.}

\item{df.abs}{A data frame with candidates' antibodies.}

\item{data}{A data frame containing demographics and medical information for
a group of waitlisted transplant candidates with color priority classification.}

\item{points.80}{A numerical value (\code{env$pt.points.minimum} - \code{env$pt.points.maximum}) for the points to a cPRA >= 80}

\item{points.50}{A numerical value (\code{env$pt.points.minimum} - \code{env$pt.points.maximum}) for the points to a cPRA >= 50}

\item{points.dialysis}{punctuaction for each month on dialysis}

\item{points.age}{A numerical value for the points to age difference}

\item{n}{A positive integer to slice the first candidates.}

\item{check.validity}{Logical to decide whether to validate input.}
}
\value{
An ordered data frame with a column 'cp' (color priority),
'sp', 'hi' and 'mmHLA'.
}
\description{
Ordering of waitlisted candidates for a given donor and
according to PT's algorithm.
}
\examples{
pt(iso = TRUE, dABO = "A",
dA = c("1","2"), dB = c("15","44"), dDR = c("1","4"),
donor.age = 65,  data = candidates,
df.abs = cabs, n = 2)
}
