% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.R
\name{pts}
\alias{pts}
\title{PT algorithm}
\usage{
pts(
  iso = TRUE,
  dABO = "O",
  dA = c("1", "2"),
  dB = c("15", "44"),
  dDR = c("1", "4"),
  donor.age = 65,
  df.abs = cabs,
  data = candidates,
  points.80 = 8,
  points.50 = 4,
  points.dialysis = 0.1,
  points.age = 4,
  n = 2,
  check.validity = TRUE
)
}
\arguments{
\item{iso}{A logical value for isogroup compatibility.}

\item{dABO}{A character value with ABO blood group (O, A, B, AB).}

\item{dA}{A two elements character vector donor's HLA-A typing.}

\item{dB}{A two elements character vector donor's HLA-B typing.}

\item{dDR}{A two elements character vector donor's HLA-DR typing.}

\item{donor.age}{A numeric value with donor's age.}

\item{df.abs}{A data frame with candidates' antibodies.}

\item{data}{A data frame containing demographics and medical information for
a group of waitlisted transplant candidates.}

\item{points.80}{A numerical value (0 - 20) for the points to a cPRA >= 80}

\item{points.50}{A numerical value (0 - 20) for the points to a cPRA >= 50}

\item{points.dialysis}{punctuation for each month on dialysis}

\item{points.age}{A numerical value for the points to age difference}

\item{n}{A positive integer to slice the first candidates.}

\item{check.validity}{Logical to decide whether to validate input.}
}
\value{
An ordered data frame with column, \code{ptsPT},
\code{SP}, \code{HI}, \code{mmHLA}, ....
}
\description{
Applies Portuguese (PT) algorithm on deceased donor's kidney allocation for transplantation.
Ordering of waitlisted candidates for a given donor according to PT's algorithm.
}
\examples{
pts(iso = TRUE, dABO = "A",
dA = c("1","2"), dB = c("15","44"), dDR = c("1","4"),
donor.age = 65,  data = candidates,
df.abs = cabs, n = 2)
}
