% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.R
\name{pts_HLA}
\alias{pts_HLA}
\title{Points for HLA mismatches}
\usage{
pts_HLA(
  itemA = 12,
  itemB = 8,
  itemC = 4,
  itemD = 2,
  itemE = 1,
  mm.A = 0,
  mm.B = 0,
  mm.DR = 0
)
}
\arguments{
\item{itemA}{Points for HLA fullmatch (no mm for HLA-A, B and DR)}

\item{itemB}{Points without mm for HLA-B and DR}

\item{itemC}{Points with 1 mm for HLA-B and DR}

\item{itemD}{Points with 1 mm for HLA-B and 1 mm for DR}

\item{itemE}{Points for remaing possibilities}

\item{mm.A}{Number of HLA-A mismatchs(0 to 2)}

\item{mm.B}{Number of HLA-B mismatchs(0 to 2)}

\item{mm.DR}{Number of HLA-DR mismatchs(0 to 2)}
}
\value{
A numerical value for pre-defined points
}
\description{
Punctuation given according to HLA mismatchs (mm) for item A) to E) from PT's algorithm
}
\examples{
pts_HLA(itemA = 12, itemB = 8, itemC = 4, itemD = 2, itemE = 1
, mm.A = 0, mm.B = 0, mm.DR = 0)
}
