% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.R
\name{pts_PRA}
\alias{pts_PRA}
\title{Points for cPRA sensitization}
\usage{
pts_PRA(cPRA = 0, points.80 = 8, points.50 = 4)
}
\arguments{
\item{cPRA}{Percentual value of cPRA (0 - 100)}

\item{points.80}{A numerical value (\code{env$pt.points.minimum} - \code{env$pt.points.maximum}) for the points to a cPRA >= 80}

\item{points.50}{A numerical value (\code{env$pt.points.minimum} - \code{env$pt.points.maximum}) for the points to a cPRA >= 50}
}
\value{
A numerical value for pre-defined points
}
\description{
Punctuation given for sensitized patients according to cPRA value
}
\examples{
pts_PRA(cPRA = 0, points.80 = 8, points.50 = 4)
}
