% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk.R
\name{ric}
\alias{ric}
\title{donor-recipient Risk Index Combination}
\usage{
ric(
  DRI = "D1",
  data = candidates.uk,
  D1R1 = 1000,
  D1R2 = 700,
  D1R3 = 350,
  D1R4 = 0,
  D2R1 = 700,
  D2R2 = 1000,
  D2R3 = 500,
  D2R4 = 350,
  D3R1 = 350,
  D3R2 = 500,
  D3R3 = 1000,
  D3R4 = 700,
  D4R1 = 0,
  D4R2 = 350,
  D4R3 = 700,
  D4R4 = 1000,
  check.validity = TRUE
)
}
\arguments{
\item{DRI}{Donor RisK Index group (env$valid.dris)}

\item{data}{A data file with candidates information for UK transplant}

\item{D1R1}{A numeric value (\code{env$dirj.minimum} - \code{env$dirj.maximum}) for the combination of indexes D1R1}

\item{D1R2}{A numeric value (\code{env$dirj.minimum} - \code{env$dirj.maximum}) for the combination of indexes D1R2}

\item{D1R3}{A numeric value (\code{env$dirj.minimum} - \code{env$dirj.maximum}) for the combination of indexes D1R3}

\item{D1R4}{A numeric value (\code{env$dirj.minimum} - \code{env$dirj.maximum}) for the combination of indexes D1R4}

\item{D2R1}{A numeric value (\code{env$dirj.minimum} - \code{env$dirj.maximum}) for the combination of indexes D2R1}

\item{D2R2}{A numeric value (\code{env$dirj.minimum} - \code{env$dirj.maximum}) for the combination of indexes D2R2}

\item{D2R3}{A numeric value (\code{env$dirj.minimum} - \code{env$dirj.maximum}) for the combination of indexes D2R3}

\item{D2R4}{A numeric value (\code{env$dirj.minimum} - \code{env$dirj.maximum}) for the combination of indexes D2R4}

\item{D3R1}{A numeric value (\code{env$dirj.minimum} - \code{env$dirj.maximum}) for the combination of indexes D3R1}

\item{D3R2}{A numeric value (\code{env$dirj.minimum} - \code{env$dirj.maximum}) for the combination of indexes D3R2}

\item{D3R3}{A numeric value (\code{env$dirj.minimum} - \code{env$dirj.maximum}) for the combination of indexes D3R3}

\item{D3R4}{A numeric value (\code{env$dirj.minimum} - \code{env$dirj.maximum}) for the combination of indexes D3R4}

\item{D4R1}{A numeric value (\code{env$dirj.minimum} - \code{env$dirj.maximum}) for the combination of indexes D4R1}

\item{D4R2}{A numeric value (\code{env$dirj.minimum} - \code{env$dirj.maximum}) for the combination of indexes D4R2}

\item{D4R3}{A numeric value (\code{env$dirj.minimum} - \code{env$dirj.maximum}) for the combination of indexes D4R3}

\item{D4R4}{A numeric value (\code{env$dirj.minimum} - \code{env$dirj.maximum}) for the combination of indexes D4R4}

\item{check.validity}{Logical to decide whether to validate input.}
}
\value{
A tibble with a new column 'ric' that gives the  Risk Index Combination.
}
\description{
computes Risk Index Combination for each pair donor-recipient
}
\examples{
ric(DRI = 'D1', data = candidates.uk,
D1R1 = 1000, D1R2 = 700, D1R3 = 350, D1R4 = 0,
D2R1 = 700, D2R2 = 1000, D2R3 = 500, D2R4 = 350,
D3R1 = 350, D3R2 = 500, D3R3 = 1000, D3R4 = 700,
D4R1 = 0, D4R2 = 350, D4R3 = 700, D4R4 = 1000, check.validity = TRUE)
}
