% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seqs}
\alias{seqs}
\title{HLA-A, -B, -C amino-acid sequences in FASTA format.}
\format{
list in FASTA format representing either nucleotide sequences or amino acid (protein) sequences, in which nucleotides or amino acids are represented using single-letter codes:
\describe{
\item{X*}{HLA-X allele}
\item{name}{allele name}
\item{Annot}{annotation}
\item{class}{SeqFastadna}
...
}
}
\source{
\url{https://www.ebi.ac.uk/ipd/imgt/hla/}
}
\usage{
seqs
}
\description{
The protein sequence of exons 2 and 3 of each allele of HLA classe I genotype;
these sequences correspond to the peptide-binding domains.
}
\keyword{datasets}
