% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seqs2}
\alias{seqs2}
\title{HLA-DRB1, -DQB1 amino-acid sequences in FASTA format.}
\format{
list in FASTA format representing either nucleotide sequences or amino acid (protein) sequences, in which nucleotides or amino acids are represented using single-letter codes:
\describe{
\item{X*}{HLA-X allele}
\item{name}{allele name}
\item{Annot}{annotation}
\item{class}{SeqFastadna}
...
}
}
\source{
\url{https://sourceforge.net/projects/granthamdist/files/HLA_ClassII_CWDonly.fas/download}
}
\usage{
seqs2
}
\description{
The protein sequence of exon 2 of each allele of HLA class II genotype;
these sequences correspond to the peptide-binding domains.
}
\keyword{datasets}
