% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult.R
\name{several}
\alias{several}
\title{Runs several time the function donor_recipient_pairs() as bootstrap.}
\usage{
several(
  iteration.number = 10,
  df.donors = donors,
  df.candidates = candidates,
  df.abs = cabs,
  algorithm = lima,
  n = 0,
  seed.number = 123,
  check.validity = TRUE,
  ...
)
}
\arguments{
\item{iteration.number}{Number of times the matchability runs.}

\item{df.donors}{A data frame containing demographics and medical information
for a pool of donors. For \code{uk} algorithm must have their respective columns.}

\item{df.candidates}{A data frame containing demographics and medical information
for a group of waitlisted transplant candidates. For \code{uk} algorithm must have respective columns.}

\item{df.abs}{A data frame with candidates' antibodies.}

\item{algorithm}{The name of the function to use. Valid options are:
\code{lima}, \code{et}, \code{pts}, \code{uk} (without quotation)}

\item{n}{A positive integer to slice the first candidates.}

\item{seed.number}{Seed for new random number.
\code{seed.number} can be \code{NA} in which case no seed is applied.}

\item{check.validity}{Logical to decide whether to validate input.}

\item{...}{all the parameters used on the function algorithm}
}
\value{
Overall statistics obtained from all runs.
}
\description{
Generic function that runs the matchability between all combinations of donors and candidates.
Runs an arbitrary number of times (\code{iteration.number}) to provide statistics.
}
\examples{
\dontrun{
several(iteration.number = 10,
df.donors = donors,
df.candidates = candidates,
df.abs = cabs,
algorithm = lima,
n = 0,
seed.number = 123,
check.validity = TRUE)
}
}
