% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histof.R
\name{sp}
\alias{sp}
\title{Senior Program classification}
\usage{
sp(donor.age, candidate.age)
}
\arguments{
\item{donor.age}{A numeric value with donor's age}

\item{candidate.age}{A numeric value with candidate's age}
}
\value{
The value 1 for a candidates older than 65 with also a donor
older than 65
}
\description{
Returns 1 when candidates' belongs to Senior Program.
Prioritization for older patients for older donors, followed for young patients
for younger donors, and for last the remaining patients.
}
\examples{
sp(donor.age = 66, candidate.age = 70)
}
