% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histof.R
\name{txscore}
\alias{txscore}
\title{TRANSPLANTSCORE (Tx Score)}
\source{
\url{https://balima.shinyapps.io/scoreTx/}
}
\usage{
txscore(
  recipient.age = 20,
  recipient.race = "White",
  recipient.causeESRD = "Other",
  recipient.dialysis = 12,
  recipient.diabetes = FALSE,
  recipient.coronary = FALSE,
  recipient.albumin = 1.5,
  recipient.hemoglobin = 10,
  donor.age = 30,
  donor.diabetes = "Absence",
  donor.ECD = FALSE,
  mmHLA_A = 0,
  mmHLA_B = 0,
  mmHLA_DR = 0
)
}
\arguments{
\item{recipient.age}{A numeric value with recipient's age}

\item{recipient.race}{A character value with recipient's race from the options:
'White', 'Black', 'Hispanic', 'Other'}

\item{recipient.causeESRD}{A numeric value with recipient's cause of End-Stage Renal
Disease, with options: 'Other', 'Diabetes', 'Hypertension',
'Glomerulonephritis', 'Cystic disease'}

\item{recipient.dialysis}{A numeric value with recipient's time on dialysis (months)}

\item{recipient.diabetes}{A logical value with recipient's diabetic status}

\item{recipient.coronary}{A logical value with recipient's coronary artery disease status}

\item{recipient.albumin}{A numeric value with recipient's albumin (g/dL)}

\item{recipient.hemoglobin}{A numeric value with recipient's hemoglobin (g/dL)}

\item{donor.age}{A numeric value with donor's age}

\item{donor.diabetes}{A logical value with donor's diabetic status, with options:
'Absence', 'Presence', 'Unknown'}

\item{donor.ECD}{A logical value regarding Extended Criteria Donor}

\item{mmHLA_A}{A numeric value (0, 1, 2) with the number of HLA-A mismatchs}

\item{mmHLA_B}{A numeric value (0, 1, 2) with the number of HLA-B mismatchs}

\item{mmHLA_DR}{A numeric value (0, 1, 2) with the number of HLA-DR mismatchs}
}
\value{
5 year probability for combined outcome of mortality or graft failure
}
\description{
Returns the estimated 5-year event (mortality or graft failure
combined outcome) probability as described by Molnar, el al (2017).
}
\examples{
txscore(recipient.age = 20,
recipient.race = "White", #insurance = 0,
recipient.causeESRD = "Other",
recipient.dialysis = 12, recipient.diabetes = FALSE,
recipient.coronary = FALSE, recipient.albumin = 1.5,
recipient.hemoglobin = 10, donor.age = 30,
donor.diabetes = "Absence",
donor.ECD = FALSE, #mmHLA = "0",
mmHLA_A = 0, mmHLA_B = 0, mmHLA_DR = 0)
}
