% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqm.R
\name{uj_matx}
\alias{uj_matx}
\title{Utility - Justice matrix}
\usage{
uj_matx(max.val = 100, ratio.util = 0.1, ratio.just = 0.1)
}
\arguments{
\item{max.val}{A numerical value corresponding o the matrix maximum value.}

\item{ratio.util}{A numerical value between 0 and 0.5 corresponding to the decreasing rate for the utility criterion (by column)}

\item{ratio.just}{A numerical value between 0 and 0.5 corresponding to the decreasing rate for the justice criterion (by column)}
}
\value{
A 6 by 6 matrix with at line 1 row 1 has the maximum value and at line 6, row 6 has it's minimum.
}
\description{
Creates a matrix with scores for utility and justice criteria on organ transplantation
}
\examples{
uj_matx(max.val = 100, ratio.util = 0.1, ratio.just = 0.1)
}
