% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk.R
\name{b_blood_penalization}
\alias{b_blood_penalization}
\title{Blood group B match points}
\usage{
b_blood_penalization(dABO = "B", cABO = "O", tier = "B", pts = -1000)
}
\arguments{
\item{dABO}{A character value for donor's ABO blood group (O, A, B, AB).}

\item{cABO}{A character value for candidate's ABO blood group (O, A, B, AB).}

\item{tier}{A character value for UK transplant candidate's TIER classification (A, B)}

\item{pts}{A negative value with penalization for B candidates}
}
\value{
A numeric value.
}
\description{
Computes penalization when donor's group O and candidate's group B
}
\examples{
b_blood_penalization(dABO = "O", cABO = "B", tier = "B", pts = -1000)
}
