% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et.R
\name{et_dialysis}
\alias{et_dialysis}
\title{ET points for time on dialysis (in months)}
\usage{
et_dialysis(dialysis = 0, month = 2.78)
}
\arguments{
\item{dialysis}{A numeric value with candidate's time on dialysis, in months
(between \code{env$dialysis.minimum} and \code{env$dialysis.maximum})}

\item{month}{A numeric value with the punctuation for each month
(between env$month.points.minimum and env$month.points.maximum)}
}
\description{
Punctuation given for each month on dialysis, within ET
Kidney allocation system
}
\examples{
et_dialysis(dialysis = 100, month = 2.78)
}
