% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et.R
\name{et_mmHLA}
\alias{et_mmHLA}
\title{ET points for mmHLA}
\usage{
et_mmHLA(
  dA = c("01", "02"),
  dB = c("03", "05"),
  dDR = c("04", "06"),
  cA = c("01", "02"),
  cB = c("03", "05"),
  cDR = c("04", "06"),
  mm0 = 400,
  mm1 = 333.33,
  mm2 = 266.67,
  mm3 = 200,
  mm4 = 133.33,
  mm5 = 66.67,
  mm6 = 0
)
}
\arguments{
\item{dA}{donor's HLA-A typing}

\item{dB}{donor's HLA-B typing}

\item{dDR}{donor's HLA-DR typing}

\item{cA}{candidate's HLA-A typing}

\item{cB}{candidate's HLA-B typing}

\item{cDR}{candidate's HLA-DR typing}

\item{mm0}{A numeric value with points for 0 HLA mm on ETKAS points table}

\item{mm1}{A numeric value with points for 1 HLA mm on ETKAS points table}

\item{mm2}{A numeric value with points for 2 HLA mm on ETKAS points table}

\item{mm3}{A numeric value with points for 3 HLA mm on ETKAS points table}

\item{mm4}{A numeric value with points for 4 HLA mm on ETKAS points table}

\item{mm5}{A numeric value with points for 5 HLA mm on ETKAS points table}

\item{mm6}{A numeric value with points for 6 HLA mm on ETKAS points table}
}
\description{
Computes HLA mismatches and the respective punctuation within ET
Kidney allocation system
}
\examples{
et_mmHLA(dA = c("01","02"), dB = c("03","05"), dDR = c("04","06"),
cA = c("01","02"), cB = c("03","05"), cDR = c("04","06"),
mm0 = 400, mm1 = 333.33, mm2 = 266.67, mm3 = 200,
mm4 = 133.33, mm5 = 66.67, mm6 = 0)
}
