% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et.R
\name{et_mmp}
\alias{et_mmp}
\title{Mismatch Probability (MMP) from ETKAS}
\source{
\url{https://www.eurotransplant.org/wp-content/uploads/2020/01/H4-Kidney.pdf}
}
\usage{
et_mmp(
  data = candidates,
  hlaA = hlaApt,
  hlaB = hlaBpt,
  hlaDR = hlaDRpt,
  abo_freq = ABOpt,
  check.validity = TRUE
)
}
\arguments{
\item{data}{A data frame containing demographics and medical information for a
group of waitlisted transplant candidates.}

\item{hlaA}{A data frame with HLA-A allele frequencies}

\item{hlaB}{A data frame with HLA-B allele frequencies}

\item{hlaDR}{A data frame with HLA-DR allele frequencies}

\item{abo_freq}{A data frame with ABO blood group frequencies}

\item{check.validity}{Logical to decide whether to validate input.}
}
\description{
Mismatch Probability (MMP) is a calculation of the probability of
receiving a kidney offer with
0 and 1 broad HLA-A, -B or split DR mismatches based on 1000 kidneys offered,
taking into account AB0 blood group rules and PRA screening. Patients receive
between 0-100 MMPs
}
\examples{
et_mmp(data = candidates,
hlaA = hlaApt, hlaB = hlaBpt, hlaDR = hlaDRpt,
abo_freq = ABOpt, check.validity = TRUE)
}
