% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histof.R
\name{hiper}
\alias{hiper}
\title{Hiperimunized classification}
\usage{
hiper(cPRA = 99, cutoff = 85)
}
\arguments{
\item{cPRA}{candidate's \code{cPRA} value}

\item{cutoff}{A value to compare candidate's \code{cPRA}}
}
\value{
A logical value TRUE when \code{cPRA} >= cutoff
}
\description{
Returns candidates' hiperimunized classification (logical) according to a
cutoff value.
}
\examples{
hiper(cPRA = 99, cutoff = 85)
}
