% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histof.R
\name{hiper}
\alias{hiper}
\title{Hiperimunized classification}
\usage{
hiper(cPRA = 99, cutoff = 85)
}
\arguments{
\item{cPRA}{candidate's cPRA value}

\item{cutoff}{A value to compare candidate's cPRA}
}
\value{
A logical value T/F when cPRA >= cutoff
}
\description{
returns candidates' hiperimunized classification according to a
cutoff value
}
\examples{
hiper(cPRA = 99, cutoff = 85)
}
