% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histoc.R
\docType{package}
\name{histoc}
\alias{histoc}
\title{histoc: Histocompatibility analysis performed in kidney allocation systems for transplantation.}
\description{
The \code{histoc} package provides functions for compatibility between organ donors and transplant candidates.
}
\section{histoc functions}{

The \code{histoc} main functions: \code{lima}, \code{pts}, \code{uk}, \code{et}, \code{several}
}

