% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histoc.R
\docType{package}
\name{histoc}
\alias{histoc}
\title{histoc: A package for histocompatibility analysis performed on kidney allocation systems for transplantation}
\description{
The histoc package provides functions for compatibility between organ donors and transplant candidates.
}
\section{histoc functions}{

The histoc functions: abo, abo.uk, age.diff, b.blood, cp, et, et_mmHLA, et-dialysis,
}

