% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hlaAet}
\alias{hlaAet}
\title{HLA-A relative frequencies.}
\format{
A data frame with 27 rows and 3 variables:
\describe{
\item{A}{HLA-A allele}
\item{n}{allele counts}
\item{freq}{allele frequencies}
...
}
}
\source{
\url{https://www.eurotransplant.org/wp-content/uploads/2020/01/H4-Kidney.pdf}
}
\usage{
hlaAet
}
\description{
A dataset containing HLA-A allele relative frequencies from EuroTransplant donors used for MMP computation.
}
\keyword{datasets}
