% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hlaApt}
\alias{hlaApt}
\title{HLA-A relative frequencies.}
\format{
A data frame with 20 rows and 3 variables:
\describe{
\item{A}{HLA-A allele}
\item{n}{allele counts}
\item{freq}{allele frequencies}
...
}
}
\source{
\url{https://www.slideshare.net/balima78/lima-2013}
}
\usage{
hlaApt
}
\description{
A dataset containing HLA-A allele relative frequencies from portuguese donors used for MMP computation.
}
\keyword{datasets}
