% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hlaDRpt}
\alias{hlaDRpt}
\title{HLA-DR relative frequencies.}
\format{
A data frame with 13 rows and 3 variables:
\describe{
\item{DR}{HLA-DR allele}
\item{n}{allele counts}
\item{freq}{allele frequencies}
...
}
}
\source{
\url{https://12f11c1f-960a-f627-594d-b8ce276384f7.filesusr.com/ugd/3e838e_dc548dede99a4db5869c3d2c20c2d16f.pdf?index=true}
}
\usage{
hlaDRpt
}
\description{
A dataset containing HLA-DR allele relative frequencies from portuguese donors used for MMP computation.
}
\keyword{datasets}
