% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt.R
\name{pts_PRA}
\alias{pts_PRA}
\title{Points for cPRA sensitization}
\usage{
pts_PRA(cPRA = 0, pts.80 = 8, pts.50 = 4)
}
\arguments{
\item{cPRA}{A numerical value (0-100) for cPRA}

\item{pts.80}{A numerical value for the points to a cPRA >= 80}

\item{pts.50}{A numerical value for the points to a cPRA >= 50}
}
\value{
A numerical value for pre-defined points
}
\description{
Punctuation given for sensitized patients according to cPRA value
}
\examples{
pts_PRA(cPRA = 0, pts.80 = 8, pts.50 = 4)
}
