% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult.R
\name{several}
\alias{several}
\title{Runs several time the function donor_recipient_pairs() for as bootstrap.}
\usage{
several(
  iteration.number = 10,
  df.donors = donors,
  df.candidates = candidates,
  df.abs = cabs,
  algorithm = lima,
  n = 0,
  seed.number = 123,
  check.validity = TRUE,
  ...
)
}
\arguments{
\item{iteration.number}{Number of times the matchability runs.}

\item{df.donors}{A data frame containing demographics and medical information
for a poll of donors. For uk algorithm must have respective columns.}

\item{df.candidates}{A data frame containing demographics and medical information
for a group of waitlisted transplant candidates. For uk algorithm must have respective columns.}

\item{df.abs}{A data frame with candidates' antibodies.}

\item{algorithm}{The name of the function to use. Valid options are: lima, et, pt, uk (without quotation)}

\item{n}{A positive integer to slice the first candidates.}

\item{seed.number}{Seed for new random number. seed.number can be NA in which case no seed is applied.}

\item{check.validity}{Logical to decide whether to validate input.}

\item{...}{all the parameters used on the function algorithm}
}
\value{
Statistics related to all the times the function ran.
}
\description{
Generic function that runs the matchability between all combinations of donors and candidates.
Runs an arbitrary number of times ('iteration.number') to provide statistics.
}
\examples{
several(iteration.number = 10,
df.donors = donors,
df.candidates = candidates,
df.abs = cabs,
algorithm = lima,
n = 0,
seed.number = 123,
check.validity = TRUE)
}
