% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uk.R
\name{uk}
\alias{uk}
\title{resume function for UK algorithm punctuation}
\usage{
uk(
  DRI = "D1",
  dA = c("1", "2"),
  dB = c("15", "44"),
  dDR = c("1", "4"),
  dABO = "O",
  donor.age = 65,
  data = candidates.uk,
  D1R1 = 1000,
  D1R2 = 700,
  D1R3 = 350,
  D1R4 = 0,
  D2R1 = 700,
  D2R2 = 1000,
  D2R3 = 500,
  D2R4 = 350,
  D3R1 = 350,
  D3R2 = 500,
  D3R3 = 1000,
  D3R4 = 700,
  D4R1 = 0,
  D4R2 = 350,
  D4R3 = 700,
  D4R4 = 1000,
  ptsDial = 1,
  a1 = 2300,
  a2 = 1500,
  b1 = 1200,
  b2 = 750,
  b3 = 400,
  m = 40,
  nn = 4.5,
  o = 4.7,
  mm1 = -100,
  mm23 = -150,
  mm46 = -250,
  pts = -1000,
  df.abs = cabs,
  n = 2,
  check.validity = TRUE
)
}
\arguments{
\item{DRI}{Donor RisK Index group (env$valid.dris)}

\item{dA}{donor's HLA-A typing.}

\item{dB}{donor's HLA-B typing.}

\item{dDR}{donor's HLA-DR typing.}

\item{dABO}{A character value with ABO blood group (\code{env$valid.blood.groups}).}

\item{donor.age}{A numeric value with donor's age.}

\item{data}{A data frame containing demographics and medical information for
a group of waitlisted transplant for UK transplant.}

\item{D1R1}{A numeric value (env$dirj.minimum - env$dirj.maximum) for the combination of indexes DiRj}

\item{D1R2}{A numeric value (env$dirj.minimum - env$dirj.maximum) for the combination of indexes DiRj}

\item{D1R3}{A numeric value (env$dirj.minimum - env$dirj.maximum) for the combination of indexes DiRj}

\item{D1R4}{A numeric value (env$dirj.minimum - env$dirj.maximum) for the combination of indexes DiRj}

\item{D2R1}{A numeric value (env$dirj.minimum - env$dirj.maximum) for the combination of indexes DiRj}

\item{D2R2}{A numeric value (env$dirj.minimum - env$dirj.maximum) for the combination of indexes DiRj}

\item{D2R3}{A numeric value (env$dirj.minimum - env$dirj.maximum) for the combination of indexes DiRj}

\item{D2R4}{A numeric value (env$dirj.minimum - env$dirj.maximum) for the combination of indexes DiRj}

\item{D3R1}{A numeric value (env$dirj.minimum - env$dirj.maximum) for the combination of indexes DiRj}

\item{D3R2}{A numeric value (env$dirj.minimum - env$dirj.maximum) for the combination of indexes DiRj}

\item{D3R3}{A numeric value (env$dirj.minimum - env$dirj.maximum) for the combination of indexes DiRj}

\item{D3R4}{A numeric value (env$dirj.minimum - env$dirj.maximum) for the combination of indexes DiRj}

\item{D4R1}{A numeric value (env$dirj.minimum - env$dirj.maximum) for the combination of indexes DiRj}

\item{D4R2}{A numeric value (env$dirj.minimum - env$dirj.maximum) for the combination of indexes DiRj}

\item{D4R3}{A numeric value (env$dirj.minimum - env$dirj.maximum) for the combination of indexes DiRj}

\item{D4R4}{A numeric value (env$dirj.minimum - env$dirj.maximum) for the combination of indexes DiRj}

\item{ptsDial}{A numeric value for the points corresponding to each month
on dialysis}

\item{a1}{A numeric value for HLA match and age combined formula:
b1*cos(age / env$adulthood.age)+a1}

\item{a2}{A numeric value for HLA match and age combined formula:
b2*cos(age / env$adulthood.age)+a2}

\item{b1}{A numeric value for HLA match and age combined formula:
b1*cos(age / env$adulthood.age)+a1}

\item{b2}{A numeric value for HLA match and age combined formula:
b2*cos(age / env$adulthood.age)+a2}

\item{b3}{A numeric value for HLA match and age combined formula:
b3*sin(age / 50)}

\item{m}{A numeric value for matchability formula: m * (1 + (MS / nn) ^ o)}

\item{nn}{A numeric value for matchability formula: m * (1 + (MS / nn) ^ o)}

\item{o}{A numeric value for matchability formula: m * (1 + (MS / nn) ^ o)}

\item{mm1}{A numeric value to penalize 1 mm}

\item{mm23}{A numeric value to penalize 2-3 mm}

\item{mm46}{A numeric value to penalize 4-6 mm}

\item{pts}{A negative value with penalization for B candidates}

\item{df.abs}{A data frame with candidates' antibodies.}

\item{n}{A positive integer to slice the first candidates.}

\item{check.validity}{Logical to decide whether to validate input.}
}
\description{
Ordering of waitlisted candidates for a given donor and according
to UK transplant algorithm.
}
\examples{
uk(DRI = 'D1', dA = c("1","2"), dB = c("15","44"), dDR = c("1","4"),
dABO = "O", donor.age = 65, data = candidates.uk,
D1R1 = 1000, D1R2 = 700, D1R3 = 350, D1R4 = 0,
D2R1 = 700, D2R2 = 1000, D2R3 = 500, D2R4 = 350,
D3R1 = 350, D3R2 = 500, D3R3 = 1000, D3R4 = 700,
D4R1 = 0, D4R2 = 350, D4R3 = 700, D4R4 = 1000,
ptsDial = 1,
a1 = 2300,  a2 = 1500, b1 = 1200, b2 = 750, b3 = 400,
m = 40, nn = 4.5, o = 4.7,
mm1 = -100, mm23 = -150, mm46 = -250,
pts = -1000,
df.abs = cabs, n = 2, check.validity = TRUE)
}
