% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histof.R
\name{xmatch}
\alias{xmatch}
\title{virtual crossmatch (XM)}
\usage{
xmatch(dA, dB, dDR, df.abs)
}
\arguments{
\item{dA}{donor's HLA-A typing}

\item{dB}{donor's HLA-B typing}

\item{dDR}{donor's HLA-DR typing}

\item{df.abs}{data frame with candidates' antibodies}
}
\value{
A dataframe with candidates' ID and XM result POS/NEG
}
\description{
returns candidates' virtual crossmatch againts donor's HLA typing
}
\examples{
xmatch(dA = c('01','02'), dB = c('05','07'), dDR = c('01','04'), df.abs = abs)
}
